% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivalvePBDB.R
\docType{data}
\name{BivalvePBDB}
\alias{BivalvePBDB}
\title{Occurrence data on bivalves from the Ypresian
Data from the Paleobiology Database via the velociraptr package on May 17th, 2017}
\format{A csv file}
\usage{
data(BivalvePBDB)
}
\description{
Occurrence data on bivalves from the Ypresian
Data from the Paleobiology Database via the velociraptr package on May 17th, 2017
}
\examples{
data(BivalvePBDB)
}
\keyword{dataset}
