% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg2rad.R
\name{deg2rad}
\alias{deg2rad}
\title{Converts degrees to radians}
\usage{
deg2rad(deg)
}
\arguments{
\item{deg}{- decimal degree to be converted to radians}
}
\value{
Returns the degree in radians
}
\description{
Converts degrees to radians
}
\examples{
deg2rad(45)
}
\references{
[1]Originally from http://www.r-bloggers.com/great-circle-distance-calculations-in-r/
}
