% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HorseShoeArea.R
\name{HorseShoeArea}
\alias{HorseShoeArea}
\title{Function to calcuate the area of a given horseshoe shape}
\usage{
HorseShoeArea(HorseShoeShape)
}
\arguments{
\item{HorseShoeShape}{- Object containing the outline of a horseshoe shape, output from PtsAlgHorseShoe function}
}
\value{
Returns the area of the horseshoe shape with the circular and rectangular portions separated
}
\description{
Function to calcuate the area of a given horseshoe shape
}
\note{
Currently forces use of a height 5/4 width and r2=2*r1 of origin (center)
}
\examples{
HorseShoeTest<-PtsAlgHorseShoe(z=2000,spacing=1,endAngles=c(-90,90))
HorseShoeArea(HorseShoeTest)
}
\references{
[1] From http://mathforum.org/library/drmath/view/51816.html
}
