% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWMatrix.R
\name{PWMatrix}
\alias{PWMatrix}
\title{Creates a sparse pairwise distance matrix of a coordinate set}
\usage{
PWMatrix(Coords)
}
\arguments{
\item{Coords}{- Two-dimensional array of longitudinal and latitudinal coordinates output from CoordCollapse() function}
}
\value{
Returns a sparse pairwise distance matrix of great circle distances between pairs of points
}
\description{
Creates a sparse pairwise distance matrix of a coordinate set
}
\examples{
longs<-runif(10,-22,33)
lats<-runif(10,-22,33)
Coords<-CoordCollapse(longs,lats)
PWMatrix(Coords)
}
